--- co-make/src/store/actions/upvoteIssueAct.js
@@ -1,5 +1,0 @@
-const UPVOTE_ISSUE = 'UPVOTE_ISSUE'
-
-export const upvoteIssue = () => (dispatch) => {
-   dispatch({ type : UPVOTE_ISSUE})
-}
