+++ co-make/src/store/reducers/userReducer.js
@@ -1,0 +1,6 @@
+export const userReducer = (state = null, action) => {
+   switch (action.type) {
+      default:
+         return state;
+   }
+}
